Fix crashing Inkscape when autotrace not installed
GUI version crashed Inkscape (!) when autotrace was not installed in Inkscape 0.91 / Ubuntu 14.04. Commented out the offending part and replaced it by a check that works for me (cannot test on another system). 

Also fixed a crash when the image did not have x and y attributes, as imported images in 0.91 only seem to have a transform attribute, but no x and y. Could be fixed better by either getting their position by calling the commandline, or by using computeBBox from simpletransform.